redaction

+++ trunk/ii02810/task_02/test/Test.cpp
@@ -44,7 +44,7 @@ TEST(UnLinemodelTest, BasicCase) {
     auto result = UnLinemodel(p, y, y0);
     ASSERT_EQ(result.size(), 1);
 
-    double expected = p.a * y - p.b * y0 * y0 + p.c * p.u + p.d * sin(p.u);
+    float expected = p.a * y - p.b * y0 * y0 + p.c * p.u + p.d * std::sinf(p.u);
     EXPECT_NEAR(result[0], expected, 1e-5);
 }
